# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/products":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["products"]},"get":{"operationId":"ProductsController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"minPrice","required":true,"in":"query","schema":{"type":"number"}},{"name":"maxPrice","required":true,"in":"query","schema":{"type":"number"}},{"name":"sortBy","required":true,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]},"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["products"]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/types":{"post":{"operationId":"TypesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTypeDto"}}}},"responses":{"201":{"description":""}},"tags":["types"]},"get":{"operationId":"TypesController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["types"]}},"/types/{id}":{"get":{"operationId":"TypesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["types"]},"patch":{"operationId":"TypesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTypeDto"}}}},"responses":{"200":{"description":""}},"tags":["types"]},"delete":{"operationId":"TypesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["types"]}},"/images":{"post":{"operationId":"ImagesController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["images"]}},"/carts":{"post":{"operationId":"CartsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCartDto"}}}},"responses":{"201":{"description":""}},"tags":["carts"]},"get":{"operationId":"CartsController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["carts"]}},"/carts/{id}":{"get":{"operationId":"CartsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["carts"]},"patch":{"operationId":"CartsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCartDto"}}}},"responses":{"200":{"description":""}},"tags":["carts"]},"delete":{"operationId":"CartsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["carts"]}},"/carts/{id}/clear":{"post":{"operationId":"CartsController_clear","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["carts"]}},"/orders":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["orders"]},"get":{"operationId":"OrdersController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/orders/{id}":{"get":{"operationId":"OrdersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]},"patch":{"operationId":"OrdersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":""}},"tags":["orders"]},"delete":{"operationId":"OrdersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/carts-products":{"post":{"operationId":"CartsProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCartsProductDto"}}}},"responses":{"201":{"description":""}},"tags":["cart_products"]},"get":{"operationId":"CartsProductsController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["cart_products"]},"delete":{"operationId":"CartsProductsController_remove","parameters":[{"name":"product_id","required":true,"in":"query","schema":{"type":"number"}},{"name":"user_id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["cart_products"]}},"/carts-products/{id}":{"get":{"operationId":"CartsProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["cart_products"]},"patch":{"operationId":"CartsProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCartsProductDto"}}}},"responses":{"200":{"description":""}},"tags":["cart_products"]}},"/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["auth"]}},"/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":""},"201":{"description":""}},"tags":["auth"]}},"/auth/sign-out":{"post":{"operationId":"AuthController_signOut","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/session":{"get":{"operationId":"AuthController_getSesssionInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["auth"]}}},"info":{"title":"","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@mail.com"},"hash":{"type":"string","example":"1"},"salt":{"type":"string","example":"1"},"surname":{"type":"string","example":"Иванов"},"name":{"type":"string","example":"Иван"},"lastname":{"type":"string","example":"Иванович"}},"required":["email","hash","salt","surname","name","lastname"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@mail.com"},"hash":{"type":"string","example":"1"},"salt":{"type":"string","example":"1"},"surname":{"type":"string","example":"Иванов"},"name":{"type":"string","example":"Иван"},"lastname":{"type":"string","example":"Иванович"}}},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","example":"Santo Stefano Vacanze romane"},"description":{"type":"string","example":"Это вино было создано богами в древнем риме когда фронтенд писали на razor asp.net"},"characteristics":{"example":[{"key":"Страна","value":"Италия"},{"key":"Артикул","value":"Santo Stefano"}],"type":"array","items":{"type":"string"}},"price":{"type":"number","example":499},"image_id":{"type":"number","example":1},"type_id":{"type":"number","example":1}},"required":["name","description","characteristics","price","image_id","type_id"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string","example":"Santo Stefano Vacanze romane"},"description":{"type":"string","example":"Это вино было создано богами в древнем риме когда фронтенд писали на razor asp.net"},"characteristics":{"example":[{"key":"Страна","value":"Италия"},{"key":"Артикул","value":"Santo Stefano"}],"type":"array","items":{"type":"string"}},"price":{"type":"number","example":499},"image_id":{"type":"number","example":1},"type_id":{"type":"number","example":1}}},"CreateTypeDto":{"type":"object","properties":{"name":{"type":"string","example":"Вино"},"image_id":{"type":"number","example":1},"url":{"type":"string","example":"vines"}},"required":["name","image_id","url"]},"UpdateTypeDto":{"type":"object","properties":{"name":{"type":"string","example":"Вино"},"image_id":{"type":"number","example":1},"url":{"type":"string","example":"vines"}}},"CreateCartDto":{"type":"object","properties":{"user_id":{"type":"number","example":1}},"required":["user_id"]},"UpdateCartDto":{"type":"object","properties":{"user_id":{"type":"number","example":1}}},"CreateOrderDto":{"type":"object","properties":{"user_id":{"type":"number","example":1},"products_id":{"type":"number","example":1},"quantity":{"type":"number","example":1},"summary":{"type":"number","example":1}},"required":["user_id","products_id","quantity","summary"]},"UpdateOrderDto":{"type":"object","properties":{"user_id":{"type":"number","example":1},"products_id":{"type":"number","example":1},"quantity":{"type":"number","example":1},"summary":{"type":"number","example":1}}},"CreateCartsProductDto":{"type":"object","properties":{"cart_id":{"type":"number","example":1},"product_id":{"type":"number","example":1},"quantity":{"type":"number","example":1}},"required":["cart_id","product_id","quantity"]},"UpdateCartsProductDto":{"type":"object","properties":{"cart_id":{"type":"number","example":1},"product_id":{"type":"number","example":1},"quantity":{"type":"number","example":1}}},"SignUpDto":{"type":"object","properties":{"email":{"type":"string","example":"user@mail.com"},"password":{"type":"string","example":"12421242"},"surname":{"type":"string","example":"Иванов"},"name":{"type":"string","example":"Иван"},"lastname":{"type":"string","example":"Иванович"}},"required":["email","password","surname","name","lastname"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"email":{"type":"string","example":"user@mail.com"},"surname":{"type":"string","example":"Чайковский"},"name":{"type":"string","example":"Василий"},"lastname":{"type":"string","example":"Янович"},"role":{"type":"string","example":"user"}},"required":["id","email","surname","name","lastname","role"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","example":"user@mail.com"},"password":{"type":"string","example":"12421242"}},"required":["email","password"]}}}}